Fix drawing characters

+++ src/kernel/font.c
@@ -9,7 +9,7 @@ void draw_char(u32 x, u32 y, u8 c) {
         // assert(0 <= y && y+FontCharHeight <= SCREEN_HEIGHT);
         for (u32 i = 0; i < FontCharHeight; ++i) {
             for (u32 j = 0; j < FontCharWidth / 8; ++j) {
-                u8 charOffset = (c - ASCII_PRINTABLE_MIN) * FontCharHeight * FontCharWidth / 8;
+                u32 charOffset = (c - ASCII_PRINTABLE_MIN) * FontCharHeight * FontCharWidth / 8;
                 u8 b = FontBitmap[charOffset + i * FontCharWidth / 8 + j];
                 for (u32 s = 0; s < 8; ++s) {
                     if (b & (1U << (7 - s)))
